Move session slice inside redux folder
